+++ src/time.rs
@@ -22,7 +22,7 @@ impl TimeContext {
         fps_tracker.resize(200, 1.0 / 60.0);
 
         TimeContext {
-            tick_rate: f64_to_duration(tick_rate),
+            tick_rate: Duration::from_secs_f64(tick_rate),
             fps_tracker,
             last_time: Instant::now(),
             lag: Duration::from_secs(0),
@@ -44,7 +44,7 @@ pub(crate) fn tick(ctx: &mut Context) {
     // Since we fill the buffer when we create the context, we can cycle it
     // here and it shouldn't reallocate.
     ctx.time.fps_tracker.pop_front();
-    ctx.time.fps_tracker.push_back(duration_to_f64(elapsed));
+    ctx.time.fps_tracker.push_back(elapsed.as_secs_f64());
 }
 
 pub(crate) fn is_tick_ready(ctx: &Context) -> bool {
@@ -57,33 +57,17 @@ pub(crate) fn consume_tick(ctx: &mut Context) {
 
 // TODO: What's the proper name for the interpolation amount? NAMING AGH
 pub(crate) fn get_alpha(ctx: &Context) -> f64 {
-    duration_to_f64(ctx.time.lag) / duration_to_f64(ctx.time.tick_rate)
-}
-
-/// Converts a `std::time::Duration` to an `f64`. This is less accurate, but
-/// usually more useful.
-pub fn duration_to_f64(duration: Duration) -> f64 {
-    let seconds = duration.as_secs() as f64;
-    let nanos = f64::from(duration.subsec_nanos()) * 1e-9;
-    seconds + nanos
-}
-
-/// Converts an `f64` to a `std::time::Duration`.
-pub fn f64_to_duration(duration: f64) -> Duration {
-    debug_assert!(duration >= 0.0);
-    let seconds = duration.trunc() as u64;
-    let nanos = (duration.fract() * 1e9) as u32;
-    Duration::new(seconds, nanos)
+    ctx.time.lag.as_secs_f64() / ctx.time.tick_rate.as_secs_f64()
 }
 
 /// Gets the update tick rate of the application, in ticks per second.
 pub fn get_tick_rate(ctx: &Context) -> f64 {
-    1.0 / duration_to_f64(ctx.time.tick_rate)
+    1.0 / ctx.time.tick_rate.as_secs_f64()
 }
 
 /// Sets the update tick rate of the application, in ticks per second.
 pub fn set_tick_rate(ctx: &mut Context, tick_rate: f64) {
-    ctx.time.tick_rate = f64_to_duration(1.0 / tick_rate);
+    ctx.time.tick_rate = Duration::from_secs_f64(1.0 / tick_rate);
 }
 
 /// Returns the current frame rate, averaged out over the last 200 frames.